Minilyrix - Development screen test
Minilyrix is a program inspired by the Minilyrics software available as a plugin to be used along with the media players in Windows systems. This program, as opposed to showing the lyrics in the top portion of the screen or anywhere in that matter, will be showing the lyrics in a ncurses based window opened along with a text based player, for example cmus.