Torrent Finder
TorrentFinder.sh is a "simple" bash script that uses wget, grep, etc. to list the top 5 torrents found on each of the 8 sites based on the search you made. tested and working on archlinux TorrentFinder should work on any other linux distro, i've seen problems when using OSX. I have tested it in cygwin and it's working just fine